async tests [nfc]: Put BackoffMachine tests where they belong.
Follow and delete a code comment at the top of
`backoffMachine-test`, suggesting that we move these tests.
